[deepspeed-deco] Refactoring code. - Unfinished - Ran black
