+++ metaflow_extensions/deepspeed/plugins/az_store.py
@@ -6,7 +6,9 @@ from typing import List, Optional, Union, Tuple
 
 from metaflow import FlowSpec
 from metaflow.metaflow_config import DATATOOLS_AZUREROOT, DATASTORE_SYSROOT_AZURE
-from metaflow.plugins.azure.blob_service_client_factory import get_azure_blob_service_client
+from metaflow.plugins.azure.blob_service_client_factory import (
+    get_azure_blob_service_client,
+)
 from metaflow.plugins.datastores.azure_storage import AzureStorage, _AzureRootClient
 from metaflow.exception import MetaflowException
 
@@ -14,8 +16,10 @@ from metaflow.exception import MetaflowException
 AzureObject = namedtuple("AzureObject", "blob url key text")
 AzureListPathResult = namedtuple("AzureListPathResult", "url key")
 
+
 class AzureDatastoreNotFoundError(MetaflowException):
     headline = "Azure Datastore Object Not Found"
+
     def __init__(self, url):
         super(AzureDatastoreNotFoundError, self).__init__(
             "The Azure Datastore object at '%s' was not found." % url
@@ -36,21 +40,26 @@ class AzureBlob(object):
     """
 
     def __init__(
-        self, 
+        self,
         run: Optional[Union[FlowSpec, "metaflow.Run"]] = None,
         blob_path: Optional[str] = None,
         run_pathspec: Optional[str] = None,
-        step_name: Optional[str] = None
+        step_name: Optional[str] = None,
     ):
-
         if blob_path is None and run_pathspec is None:
-            raise ValueError("Either blob_path or run_pathspec must be specified in the AzureBlob instantiation.")
+            raise ValueError(
+                "Either blob_path or run_pathspec must be specified in the AzureBlob instantiation."
+            )
         if blob_path and run_pathspec:
-            raise ValueError("Cannot specify both blob_path and run_pathspec in the AzureBlob instantiation.")
+            raise ValueError(
+                "Cannot specify both blob_path and run_pathspec in the AzureBlob instantiation."
+            )
 
         # TODO: tmp
-        if blob_path: 
-            raise NotImplementedError("Custom blob_path is not yet implemented in AzureBlob. Pass flow_name and run_id instead.")
+        if blob_path:
+            raise NotImplementedError(
+                "Custom blob_path is not yet implemented in AzureBlob. Pass flow_name and run_id instead."
+            )
 
         self._datastore = AzureStorage()
         self._datastore_root = DATASTORE_SYSROOT_AZURE
@@ -61,9 +70,7 @@ class AzureBlob(object):
 
     def get_datastore_key_location(self, key: Optional[str] = None):
         if self._step_name is None:
-            return os.path.join(
-                self._run_pathspec, key if key else ""
-            ).rstrip("/")
+            return os.path.join(self._run_pathspec, key if key else "").rstrip("/")
         else:
             return os.path.join(
                 self._run_pathspec, self._step_name, key if key else ""
@@ -97,14 +104,14 @@ class AzureBlob(object):
                                 blob=blob_bytes,
                                 url=datastore_url,
                                 text=blob_bytes.decode("utf-8"),
-                                key=datastore_url[len(keyless_root):].strip("/")
+                                key=datastore_url[len(keyless_root) :].strip("/"),
                             )
                         except UnicodeDecodeError:
                             return AzureObject(
                                 blob=blob_bytes,
                                 url=datastore_url,
                                 text=None,
-                                key=datastore_url[len(keyless_root):].strip("/")
+                                key=datastore_url[len(keyless_root) :].strip("/"),
                             )
                 else:
                     raise AzureDatastoreNotFoundError(datastore_url)
@@ -118,7 +125,7 @@ class AzureBlob(object):
             keys = [self.get_datastore_key_location(key) for key in keys]
         results = []
         for list_content_result in self._datastore.list_content(keys):
-            key_suffix = list_content_result.path[len(keyless_root):].strip("/")
+            key_suffix = list_content_result.path[len(keyless_root) :].strip("/")
             if list_content_result.is_file:
                 results.append(
                     AzureListPathResult(url=list_content_result.path, key=key_suffix)
@@ -132,17 +139,14 @@ class AzureBlob(object):
         for key, path in key_paths:
             with open(path, "rb") as f:
                 self.put(key, f.read(), overwrite=overwrite)
-            results.append(
-                self.get_datastore_key_location(key)
-            )
+            results.append(self.get_datastore_key_location(key))
         return results
 
     def get_files(self, key_paths: List[Tuple[str, str]], as_binary=False):
         for key, path in key_paths:
-
             if not os.path.exists(os.path.dirname(path)):
                 os.makedirs(os.path.dirname(path))
-                
+
             if as_binary:
                 with open(path, "wb") as f:
                     f.write(self.get(key).blob)
