+++ metaflow_extensions/deepspeed/plugins/constants.py
@@ -1,0 +1,8 @@
+HOSTFILE_IP_KEY = "hostfile_ips"
+HOSTFILE = "hostfile.txt"
+DEEPSPEED_JOB_COMPLETE_VAR = "mpi_job_status"
+PUBLIC_KEY_RECEIVED_VAR = "public_key_received"
+CONTROL_TASK_DONE_PATH = "control"
+MPI_PUBLIC_KEY_PATH = "public_keys"
+DEEPSPEED_ENV_FILE = ".deepspeed_env"  # https://github.com/microsoft/DeepSpeed/blob/24f20ef0a105d32f6085fe0d3b1c2f9324a6262c/docs/_tutorials/getting-started.md?plain=1#L230-L254
+DEEPSPEED_SUFFIX = "mf.deepspeed_datastore"
